switch to long double to increase integer precision
Since double's tend to have 53-bits of precision, doing 64-bit integer
math can quickly (and perniciously) go wrong.  Switch to long doubles
everywhere to fix that on most current platforms -- the standard allows
for double & long double to be equal, but 64-bit systems are basically
guaranteed to have higher precision.

This is a bandaid until pcalc can be switched to an arbitrary precision
math library like GMP.

Fixes #11.

+++ hocdecl.h
@@ -24,24 +24,24 @@ extern void init( void) ;
 **      functions in MATH.C
 */
 
-extern  double Log(double x);
-extern  double Log10(double x);
-extern  double Exp(double x);
-extern  double Pow(double x,double y);
-extern  double integer(double x);
-extern  double Sqrt(double x);
-extern  double f2c(double d);
-extern  double c2f(double d);
-extern  double in2mm(double d);
-extern  double mm2in(double d);
-extern  double po2kg(double d);
-extern  double kg2po(double d);
+extern  long double Log(long double x);
+extern  long double Log10(long double x);
+extern  long double Exp(long double x);
+extern  long double Pow(long double x, long double y);
+extern  long double integer(long double x);
+extern  long double Sqrt(long double x);
+extern  long double f2c(long double d);
+extern  long double c2f(long double d);
+extern  long double in2mm(long double d);
+extern  long double mm2in(long double d);
+extern  long double po2kg(long double d);
+extern  long double kg2po(long double d);
 
 /*
 **      functions in FUNCT.C
 */
 
-extern  int print(double);
+extern  int print(long double);
 extern  int echo(const char *str);
 extern  int echo_nl(const char *str);
 extern  int ddate(const char *str);
